pyproject.toml
[tool.poetry]
name = "wb-hometask-11"
version = "0.1.0"
description = ""
authors = ["Misha-HF <144258555+Misha-HF@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.0"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
pydantic = "^2.7.0"
email-validator = "^2.1.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
pydentic = "^0.0.1.dev3"
datetime = "^5.5"
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.1"
redis = "^5.0.4"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.40.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
pytest = "^8.2.0"
httpx = "^0.27.0"

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
